task_app/requirements.txt
Django==3.1.7
djangorestframework==3.12.2
django-allauth==0.44.0
django-rest-auth==0.9.5